feat: Prevent a balance from being deleted if a payment is linked to it.

+++ src/use-cases/balances-use-case/delete-balances.ts
@@ -1,5 +1,7 @@
 import { Saldos } from "@prisma/client";
 import { BalancesRepository } from "../../repositories/balance-repository";
+import { PaymentsRepository } from "../../repositories/payments-repository";
+import { BalanceWithPaymentError } from "../errors/balance-with-payment";
 
 interface DeleteBalancesUseCaseRequest {
   balanceId: string;
@@ -10,11 +12,24 @@ interface DeleteBalancesUseCaseResponse {
 }
 
 export class DeleteBalancesUseCase {
-  constructor(private balancesRepository: BalancesRepository) {}
+  constructor(
+    private balancesRepository: BalancesRepository,
+    private paymentRepository: PaymentsRepository
+  ) {}
 
   async execute({
     balanceId,
   }: DeleteBalancesUseCaseRequest): Promise<DeleteBalancesUseCaseResponse> {
+    const searchPaymentWithSameBalanceId =
+      await this.paymentRepository.getPayments();
+    const searchBalanceIdinPayment = searchPaymentWithSameBalanceId.map(
+      (payment) => payment.saldos_Id
+    );
+
+    if (searchBalanceIdinPayment.includes(balanceId)) {
+      throw new BalanceWithPaymentError();
+    }
+
     const BalancesDelete = await this.balancesRepository.deleteBalances(
       balanceId
     );
